backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
boto3==1.38.28
botocore==1.38.28
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
passlib==1.7.4
Pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing-extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.0
requests==2.32.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
